Bring back animation controls and de-duplicate code
While at it, this adds some ground work for animation ordering. See #1882.
