+++ assets/src/components/index.js
@@ -1,0 +1,6 @@
+export { default as AnimationControls } from './animation-controls';
+export { default as withAmpStorySettings } from './with-amp-story-settings';
+export { default as withParentBlock } from './with-parent-block';
+export { default as withHasSelectedInnerblock } from './with-has-selected-inner-block';
+export { default as withBlockName } from './with-block-name';
+export { default as withAttributes } from './with-attributes';
