+++ assets/src/components/with-has-selected-inner-block.js
@@ -1,0 +1,18 @@
+/**
+ * WordPress dependencies
+ */
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { withSelect } from '@wordpress/data';
+
+export default createHigherOrderComponent(
+	withSelect(
+		( select, props ) => {
+			const { hasSelectedInnerBlock } = select( 'core/editor' );
+
+			return {
+				hasSelectedInnerBlock: hasSelectedInnerBlock( props.clientId, true ),
+			};
+		}
+	),
+	'withHasSelectedInnerBlock'
+);
