+++ assets/src/components/with-parent-block.js
@@ -1,0 +1,18 @@
+/**
+ * WordPress dependencies
+ */
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { withSelect } from '@wordpress/data';
+
+export default createHigherOrderComponent(
+	withSelect(
+		( select, props ) => {
+			const { getBlockRootClientId, getBlock } = select( 'core/editor' );
+
+			return {
+				parentBlock: getBlock( getBlockRootClientId( props.clientId ) ),
+			};
+		}
+	),
+	'withParentBlock'
+);
