+++ webpack.config.js
@@ -72,6 +72,7 @@ module.exports = {
 		'./assets/js/amp-block-editor-toggle-compiled': './assets/src/amp-block-editor-toggle.js',
 		'./assets/js/amp-validation-detail-toggle-compiled': './assets/src/amp-validation-detail-toggle.js',
 		'./assets/js/amp-validation-single-error-url-details-compiled': './assets/src/amp-validation-single-error-url-details.js',
+		'./assets/js/amp-story-editor-blocks-compiled': './assets/src/amp-story-editor-blocks.js',
 		'./assets/js/amp-story-blocks-compiled': './assets/src/blocks/stories.js',
 	},
 	output: {
